apps/inference/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
tiktoken==0.7.0
openai==1.43.0
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
markdown==3.6
beautifulsoup4==4.12.3
rank-bm25==0.2.2
